Read files of the SWISS tournament software

readSwissTournament returns all entries of a SWISS tournament. summary returns a ranking summary of a SWISS tournament.

readSwissTournament(filename)

# S3 method for SwissTournament
summary(object, ...)

# S3 method for SwissTournament
print(x, fullReport = FALSE, ...)

# S3 method for SwissTournamentSummary
print(x, fullReport = FALSE, ...)

Arguments

filename
(character) name of the XML file returned by the program

object
a valid SwissTournament object

...
additional possible parameters passed on to other functions (not used yet)

x
a valid SwissTournament or SwissTournamentSummary object

fullReport
(logical) print a full ranking over each round or only the latest ranking

Value

readSwissTournament returns a S3 object SwissTournament, which is implemente as a list with the following entries:

tournamentVersion (integer)

scoreCalculation (character)

rankingComparator (character)

teams (a data.frame of teams)

entryId (integer)

teamName (character)

teamCity (character)

rounds (a data.frame of all rounds and matches)

round (integer)

game (interger)

entryId1 (integer)

entryId2 (integer)

points1 (integer)

points2 (integer)

finished (logical)

summary returns an object of class SwissTournamentSummary, which is a derived class of SwissTournament and contains all fields of the original SwissTorunament plus the following fields:

stats (list)

teamsParticipating (integer)

roundsPlayed (integer)

roundsSceduled (integer)

matchesPlayed (integer)

matchesScheduled (integer)

rankings (a data.frame of rankings and related values)

round (integer)

rank (integer)

entryId (integer)

opp (a list of integers) each entry indicates a opponent team thats already been played against

scoreCum (integer) cumulated game scores with respect to selected scoring function

pointsCum (integer) cumulated points scored

pointsDiffCum (integer) cumulated point differences

BHZ (integer) Buchholzzahl

Details

The SWISS tournament software is a old (2008) piece of software that allows to manage a jugger tournament as a variant of the SWISS system. It returns a XML file with all tournament information, teams, rounds, matches and results. This function reads those information into R for analysis purposes.

Examples

# NOT RUN {
# reads the tournament
hanseaticjuggercup2012 <- readSwissTournament("HanseaticJuggerCupR6")

# ad hoc information about rankings
summary(hanseaticjuggercup2012)
# }